Add test about limiting the number of returned todos
One area of the code that was not tested was the part of `TodoController` that looks for query parameter "limit" and limits the number of todos returned by the request. This test uses the same patterns are other tests of query parameters on the getTodos endpoint. One thing that is a little special here is that the number needs to be able to be used as a number or as a string, so there are a couple of places this might be helpful as a model for other tests. We needed to also make the LIMIT_KEY field of `TodoController` public so that it would be visible to the test file.